+++ server/src/user/user.router.js
@@ -4,22 +4,23 @@ const UserNotFoundException = require('./UserNotFoundException');
 const router = express.Router();
 const pagination = require('../shared/pagination');
 const idNumberControl = require('../shared/idNumberControl');
-const UserService = require('./UserService');
+const UserService = require('./user.service');
 const {body,validationResult} = require('express-validator');
 const ValidationException = require('../shared/ValidationException');
+const basicAuthentication = require('../shared/basicAuthentication');
 
 router.post('/users',
   body('username')
-    .notEmpty().withMessage('Username cannot be null')
+    .notEmpty().withMessage('username_null')
     .bail()
-    .isLength({min: 4,max: 32}).withMessage('Username must have min 4 max 32 characters'),
+    .isLength({min: 4,max: 32}).withMessage('username_size'),
   body('email')
-    .isEmail().withMessage('Must be a valid e-mail address')
+    .isEmail().withMessage('email_invalid')
     .bail()
     .custom(async (email) => {
       const user = await UserService.findByEmail(email)
       if (user) {
-        throw new Error('Email in use');
+        throw new Error('email_inuse');
       }
     })
   ,
@@ -29,7 +30,7 @@ router.post('/users',
       return next(new ValidationException(errors.array()));
     }
     await UserService.create(req.body);
-    res.send("success");
+    res.send({message: req.t("user_create_success")});
   })
 
 router.get('/users',pagination,async (req,res) => {
@@ -47,16 +48,34 @@ router.get('/users/:id',idNumberControl,async (req,res,next) => {
   }
 })
 
-router.put('/users/:id',idNumberControl,async (req,res) => {
+router.put('/users/:id',idNumberControl,basicAuthentication,async (req,res) => {
+  const authenticatedUser = req.authenticatedUser;
+  if (!authenticatedUser) {
+    return res.status(403).send({message: 'Forbidden'});
+  }
+
   const id = req.params.id;
+
+  if (authenticatedUser.id != id) {
+    return res.status(403).send({message: 'Forbidden'});
+  }
   const user = await User.findOne({where: {id: id}});
   user.username = req.body.username;
   await user.save();
   res.send('updated');
 })
 
-router.delete('/users/:id',idNumberControl,async (req,res) => {
+router.delete('/users/:id',idNumberControl,basicAuthentication,async (req,res) => {
+  const authenticatedUser = req.authenticatedUser;
+  if (!authenticatedUser) {
+    return res.status(403).send({message: 'Forbidden'});
+  }
+
   const id = req.params.id;
+
+  if (authenticatedUser.id != id) {
+    return res.status(403).send({message: 'Forbidden'});
+  }
   await User.destroy({where: {id: id}});
   res.send('removed');
 })
